Fix name of ECS task use to publish library to S3
The S3 publishing script needs to call the `s3publish-intg` ECS task to publish
the library.

This step broke recently when we split up the Jenkinsfiles for deploying the migrations
and publishing the library, because the `STAGE` parameter was omitted from
Jenkinsfile-publish-library.

Hard-code the "intg" suffix for now. In future it should be removed altogether, because
the library is published once to the management account, rather than separately to
each TDR environment: https://national-archives.atlassian.net/browse/TDR-465