TDR-1073 Remove default consignment sequence.
The consignment sequence is being incremented twice. It's being called with `select nextval('consignment_sequence_id')` and then it's being called again with the insert to the Consignment table.

If we want to keep the consignment reference mandatory, which we do, then the only option seems to be to remove the default. At no point are we inserting an empty value in here so we don't need the default.