Make series nullable
Court judgment transfer will not have a series assigned to them

To support this in the 'Consignment' table the 'SeriesId' column will be nullable

A link between the transferring body and the consignment will still need to be maintain, so a 'BodyId' column will be added to the 'Consignment' table

Currently the link between transferring body and the consignment is made through the association of the consignment's series. This will not be possible where there is no series.