TDR-2852 Update TitleAlternate description.
This ticket asks for hint text to be added to the AlternateTitle field.
In other places, we're using the description for hint text and it
doesn't look like we're using the description for this field for
anything at the moment so I think this is ok.

I'll update the frontend to use this on the page.