Use the same permissions as the real migrations user.
Because the real migration uses the migrations_user and the test is
running with admin permissions, we can have problems where the tests
pass but the run in the lambda fails because of permissions issues.

This creates a user with the same permissions as the migrations user we
use for production.